DPL: move DataProcessingContext and DeviceContext to separate header
